Initializing Multi-Module Application
This is the Initialization for the Iron Steed Motorcycle Mechanic application.

There will be a customer facing front end which allows people to interact with the website, create a profile, add a motorcycle, and schedule a time and date for which to have the motorcycle seviced.

The back end will allow Administration of the website, such as : Adding new employees, editing employee information, approving customer appointments, and some modularity with website information