JQuery import and photo styling
Extracted the JQuery from userForm that could be extracted and moved is to the user_form.js file for cleaner look.

Adding an row for user input within the user_form for a user portrait with a hidden input for the name of the existing user photo name so that the user can be saved without deleting the existing user portrait

Styling for photos displayed on user_form when picture is uploaded, and for displaying user portrait on the users list page

user_form.js adds an event listener to the  input for the image file. when a file is added it:
-validating that uploaded pictures are smaller than 1MB in size, and reject the picture/ reportValidity() if the picture is larger

-if picture is valid changes the src attribute of the image for the user portrait to the file passed to showImageThumbnail within user_form.js

adding logic for displaying the suer photo on the list of users page using th:if and th:unless to read if user.photos is null

using the Transiant annotation on the User Object Class to persist the default-user.png picture for users without a picture, or with a picture assigned but that has been deleted from the database