Exporting user information additions
Creation of an abstract class holding a super constructor for the file exporter  that takes a the content type of the exported file, and the extension and creates the name for the file from that is created. Also sets the HTTPServeleResponse content type and header based on the passed values.

Adding the links to the file export types to the users listing page.

editing the CSV exporter to take advantage of the super sontructor.

Creating an Excel exporter by taking advantage of the apaceh.poi library,