Creating Admin login page
Changed the request permissions FROM "PermitAll()" TO "Authenticated()" requiring a user to authenticate before being granted access to the Control Panel.

Adding the GetMapping to the MainController for "/login" that directs to login.html

Very basic Form on the login.html page with a REQUIRED email and password to submit the form

Creation and editing of the login.css page for stylization matching the theme of the application within the control panel.