Additinal details for update account
Adding the save details method for functionality of checking whether photo has been updated and cleaning previous photo.

adding the first name and last name to the logged user to update every time

Documentation for the function of the postSaveDetails method

updating the ISMMUserDetails to have methods for retreiving the first and last name for use within the controller.