New Categories form finishing parent and image
Adding factory methods to the Category Object to allow retrieval of just Category id and name when passed inside the copyIdAndName method for displaying the optional parent categories and the accompanying id within a drop down menu within the form.
   --These factory methods are taken advantage of within the CategoryService under listCategoriesUsedInForm() AND listChildren(List<Categories> , Category , int)

reduced the maximum size of image allowed to 100KB.

Categories list page will show a message after a new category is created, that will confirm the new category was created.