Checking for uniqueness in Categories form
Checking for Uniqueness for the Categories Name and Alias.

The method for showing the modal dialog has been changed for conciseness and readability, differentiating whether we want to show Warning modal dialog or an Error dialog, with the ability to pass a message into the method to be displayed to the user.
--The  methods have been moved into the common form JavaScript file to be used across all forms that take advantage of the modal text

The modal text itself has been moved to its own fragment for conciseness and readability among programmers.

The Post Mapping with the rest Controller has been added to the Categories page so that the form can send the information to the method to check for uniqueness before submitting the form.