Basic listing page for brands
The html page for the brand listing page has the skeleton of functionality to be edited. Navigation was changed to allow an employee other than an admin to view the brands page. will be updated in the future.

The BrandController and BrandService aid in showing the brands information to the Controller.

WebSecurity needed to be adjust to reflect who has permissions to access the brands listing page.