Creating brand form
Creating the new brand form.

All field within the form are required: name, categories, and an image for the brand.

within the Controller the Category Service is used in order to take advantage of its functionality for listing categories within a hierarchical order without having to repeat code.

the Javascript at the bottom of brand_form.html for appending a div with the selected category to the form field Chosen Categories was moved to the common_form.js. it is commented top express where its functionality is called.