Creating Product Repo Tests and Controller and Service
changed the navigation to have a link to the products listing page.

created the controller and view for listing the products.

WebSecurityConfig had been updated to Admins and mechanics can view the products at this time.

Added the service to allow refined controller to model communication

Over-rideing the toString method  on the products to shorten output