Checking for uniqueness in the Brand Form
Checking for uniqueness within the brand form when a new brand is being created, and when an existing brand is being edited.

While creating the services to check for uniqueness it was being tested within the BrandServiceTests class.

The JS method to check for uniqueness is ran when the form is submitted, a query is sent to the controller and the form submission is stalled. Mapping will return a String object stating whether its a duplicate Key for that object of if the designation is permited.

The brand constructor was added for the purpose of the BrandServiceTests