Move script specific to Brand form
Moving JS specific to the brand form back into the script at the bottom of the Brand_Form.HTML page.

Having the function within the common_form was causing naming conflicts with common form being loaded into the page. i chose to move it as to not sacrifice naming specificity and better readability.