Coding the Create New Product form's Overview tab
Providing mappings within the Product Controller for creating a new product with a list. The view is provided a list of brands and a new Product object to be edited within the form.

A Product Overview page has been created to host fragments within the products page  for the distinct tabs: Overview, Description, Images, and Details.

A Category DTO was made for use when populating the Category list within the New Product Form Overview tab as to ensure the original category object will not be altered.

A RestController has been added to the Brant controller for the Category list being queried for each individual brand.

And a BrandNotFouctRestController Exception has been added in the Scenario, where Brand is not found a default message will be provided