Adding ability to remove an element upon replacing page head
Adding the ability to remove the tag when replacing the common_page_head so that it can be integrated within other pages within providing a head tag inside another head tag.

subsequent pages were changed to reflect that the Head tag to NOT to be replaced when using the thymeleaf fragment.