Adding Product Description and Product Shipping Tab Content
add two additional html pages that will act as fragments within the product form.
A product Description tab that will provide a short and long description for each product to be displayed when information for the product is requested.
A Product Shipping page for details about the package such as Dimensions of the product (W x L x H) and a fourth field for weight.