Upload Products main image
Adding the product fragment for adding the main image during the NEW/EDITING Product form.

Integrating that fragment into the base Product Form page.

Adjusting the Product Save mapping in the ProductController to handle whether the image is present or not, and cleaning the old images before adding the new images to reduce bloat.

Lastly the Images added between users, brands, and products can very in size requirements, so the constant for max file size is adjusted within each form.